requirements.txt
PyYAML
pytest
pycodestyle
tabulate
frozendict
https://github.com/codecov/codecov-python/tarball/v2.0.6
pytest-cov
